pyscript/requirements.txt
Pillow==11.0.0
opencv-python==4.8.1.78
